+++ run_tests.sh
@@ -4,16 +4,23 @@
 
 VENV="venv"
 
+cd $(dirname "$0")
+
 if [ ! -d "$VENV" ]
 then
-  echo "Error: Could not find virtual environment '$VENV'."
-  echo "Create the environment with:\n\n    python -m venv $VENV.\n"
-  exit 1
+  echo "Warning: Could not find virtual environment '$VENV'."
+  echo "Creating virtual environment '$VENV'".
+
+  python3 -m venv "$VENV"
+  [[ "$?" != "0" ]] && "Error: Couldn't create virtual environment" && exit 1
 fi
 
 source "$VENV/bin/activate"
 
 cd app
 
+echo "Checking virtual environment dependencies"
+pip install --upgrade -r requirements.txt
+
 echo "Running python unittests"
-coverage run --omit "../$VENV/*" -m unittest test_api.py && coverage report
+coverage run --omit "../$VENV/*" -m unittest tests/test*.py && coverage report
